+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/AutoTransforms.cs
@@ -132,9 +132,9 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto
 			 new ConstantMove.Or64(),
 			 new ConstantMove.Xor32(),
 			 new ConstantMove.Xor64(),
-			 new Simplification.Move32Propagation(),
-			 new Simplification.Move64Propagation(),
-			 new Simplification.MoveObjectPropagation(),
+			 new Simplification.Move32Coalescing(),
+			 new Simplification.Move64Coalescing(),
+			 new Simplification.MoveObjectCoalescing(),
 			 new Simplification.Not32Twice(),
 			 new Simplification.Not64Twice(),
 			 new Simplification.GetLow32FromTo64(),
