+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/Simplification/Move32Coalescing.cs
@@ -2,18 +2,16 @@
 
 // This code was generated by an automated template.
 
-using Mosa.Platform.x64;
 using Mosa.Compiler.Framework;
-using Mosa.Compiler.Framework.Transforms;
 
-namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
+namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 {
 	/// <summary>
-	/// Mov64Consolidate
+	/// Move32Coalescing
 	/// </summary>
-	public sealed class Mov64Consolidate : BaseTransform
+	public sealed class Move32Coalescing : BaseTransform
 	{
-		public Mov64Consolidate() : base(X64.Mov64, TransformType.Auto | TransformType.Optimization)
+		public Move32Coalescing() : base(IRInstruction.Move32, TransformType.Auto | TransformType.Optimization)
 		{
 		}
 
@@ -25,10 +23,10 @@ namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
 			if (context.Operand1.Definitions.Count != 1)
 				return false;
 
-			if (context.Operand1.Definitions[0].Instruction != X64.Mov64)
+			if (context.Operand1.Definitions[0].Instruction != IRInstruction.Move32)
 				return false;
 
-			if (!IsVirtualRegister(context.Operand1.Definitions[0].Operand1))
+			if (IsCPURegister(context.Operand1.Definitions[0].Operand1))
 				return false;
 
 			return true;
@@ -40,7 +38,7 @@ namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
 
 			var t1 = context.Operand1.Definitions[0].Operand1;
 
-			context.SetInstruction(X64.Mov64, result, t1);
+			context.SetInstruction(IRInstruction.Move32, result, t1);
 		}
 	}
 }
