+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/Simplification/Move64Coalescing.cs
@@ -2,18 +2,16 @@
 
 // This code was generated by an automated template.
 
-using Mosa.Platform.x86;
 using Mosa.Compiler.Framework;
-using Mosa.Compiler.Framework.Transforms;
 
-namespace Mosa.Platform.x86.Transforms.Optimizations.Auto.Standard
+namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 {
 	/// <summary>
-	/// Mov32Consolidate
+	/// Move64Coalescing
 	/// </summary>
-	public sealed class Mov32Consolidate : BaseTransform
+	public sealed class Move64Coalescing : BaseTransform
 	{
-		public Mov32Consolidate() : base(X86.Mov32, TransformType.Auto | TransformType.Optimization)
+		public Move64Coalescing() : base(IRInstruction.Move64, TransformType.Auto | TransformType.Optimization)
 		{
 		}
 
@@ -25,10 +23,10 @@ namespace Mosa.Platform.x86.Transforms.Optimizations.Auto.Standard
 			if (context.Operand1.Definitions.Count != 1)
 				return false;
 
-			if (context.Operand1.Definitions[0].Instruction != X86.Mov32)
+			if (context.Operand1.Definitions[0].Instruction != IRInstruction.Move64)
 				return false;
 
-			if (!IsVirtualRegister(context.Operand1.Definitions[0].Operand1))
+			if (IsCPURegister(context.Operand1.Definitions[0].Operand1))
 				return false;
 
 			return true;
@@ -40,7 +38,7 @@ namespace Mosa.Platform.x86.Transforms.Optimizations.Auto.Standard
 
 			var t1 = context.Operand1.Definitions[0].Operand1;
 
-			context.SetInstruction(X86.Mov32, result, t1);
+			context.SetInstruction(IRInstruction.Move64, result, t1);
 		}
 	}
 }
