+++ Source/Mosa.Compiler.Framework/Transforms/Optimizations/Auto/Simplification/MoveObjectCoalescing.cs
@@ -2,18 +2,16 @@
 
 // This code was generated by an automated template.
 
-using Mosa.Platform.x64;
 using Mosa.Compiler.Framework;
-using Mosa.Compiler.Framework.Transforms;
 
-namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
+namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 {
 	/// <summary>
-	/// Mov32Consolidate
+	/// MoveObjectCoalescing
 	/// </summary>
-	public sealed class Mov32Consolidate : BaseTransform
+	public sealed class MoveObjectCoalescing : BaseTransform
 	{
-		public Mov32Consolidate() : base(X64.Mov32, TransformType.Auto | TransformType.Optimization)
+		public MoveObjectCoalescing() : base(IRInstruction.MoveObject, TransformType.Auto | TransformType.Optimization)
 		{
 		}
 
@@ -25,10 +23,7 @@ namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
 			if (context.Operand1.Definitions.Count != 1)
 				return false;
 
-			if (context.Operand1.Definitions[0].Instruction != X64.Mov32)
-				return false;
-
-			if (!IsVirtualRegister(context.Operand1.Definitions[0].Operand1))
+			if (context.Operand1.Definitions[0].Instruction != IRInstruction.MoveObject)
 				return false;
 
 			return true;
@@ -40,7 +35,7 @@ namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
 
 			var t1 = context.Operand1.Definitions[0].Operand1;
 
-			context.SetInstruction(X64.Mov32, result, t1);
+			context.SetInstruction(IRInstruction.MoveObject, result, t1);
 		}
 	}
 }
