+++ Source/Mosa.Platform.x64/Transforms/Optimizations/Auto/Standard/Mov32Coalescing.cs
@@ -2,16 +2,18 @@
 
 // This code was generated by an automated template.
 
+using Mosa.Platform.x64;
 using Mosa.Compiler.Framework;
+using Mosa.Compiler.Framework.Transforms;
 
-namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
+namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
 {
 	/// <summary>
-	/// MoveObjectPropagation
+	/// Mov32Coalescing
 	/// </summary>
-	public sealed class MoveObjectPropagation : BaseTransform
+	public sealed class Mov32Coalescing : BaseTransform
 	{
-		public MoveObjectPropagation() : base(IRInstruction.MoveObject, TransformType.Auto | TransformType.Optimization)
+		public Mov32Coalescing() : base(X64.Mov32, TransformType.Auto | TransformType.Optimization)
 		{
 		}
 
@@ -23,7 +25,10 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 			if (context.Operand1.Definitions.Count != 1)
 				return false;
 
-			if (context.Operand1.Definitions[0].Instruction != IRInstruction.MoveObject)
+			if (context.Operand1.Definitions[0].Instruction != X64.Mov32)
+				return false;
+
+			if (!IsVirtualRegister(context.Operand1.Definitions[0].Operand1))
 				return false;
 
 			return true;
@@ -35,7 +40,7 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 
 			var t1 = context.Operand1.Definitions[0].Operand1;
 
-			context.SetInstruction(IRInstruction.MoveObject, result, t1);
+			context.SetInstruction(X64.Mov32, result, t1);
 		}
 	}
 }
