+++ Source/Mosa.Platform.x64/Transforms/Optimizations/Auto/Standard/Mov64Coalescing.cs
@@ -2,16 +2,18 @@
 
 // This code was generated by an automated template.
 
+using Mosa.Platform.x64;
 using Mosa.Compiler.Framework;
+using Mosa.Compiler.Framework.Transforms;
 
-namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
+namespace Mosa.Platform.x64.Transforms.Optimizations.Auto.Standard
 {
 	/// <summary>
-	/// Move64Propagation
+	/// Mov64Coalescing
 	/// </summary>
-	public sealed class Move64Propagation : BaseTransform
+	public sealed class Mov64Coalescing : BaseTransform
 	{
-		public Move64Propagation() : base(IRInstruction.Move64, TransformType.Auto | TransformType.Optimization)
+		public Mov64Coalescing() : base(X64.Mov64, TransformType.Auto | TransformType.Optimization)
 		{
 		}
 
@@ -23,10 +25,10 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 			if (context.Operand1.Definitions.Count != 1)
 				return false;
 
-			if (context.Operand1.Definitions[0].Instruction != IRInstruction.Move64)
+			if (context.Operand1.Definitions[0].Instruction != X64.Mov64)
 				return false;
 
-			if (IsCPURegister(context.Operand1.Definitions[0].Operand1))
+			if (!IsVirtualRegister(context.Operand1.Definitions[0].Operand1))
 				return false;
 
 			return true;
@@ -38,7 +40,7 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 
 			var t1 = context.Operand1.Definitions[0].Operand1;
 
-			context.SetInstruction(IRInstruction.Move64, result, t1);
+			context.SetInstruction(X64.Mov64, result, t1);
 		}
 	}
 }
