+++ Source/Mosa.Platform.x64/Transforms/Optimizations/Manual/Special/Mov32Coalescing.cs
@@ -5,9 +5,9 @@ using Mosa.Compiler.Framework.Transforms;
 
 namespace Mosa.Platform.x64.Transforms.Optimizations.Manual.Special
 {
-	public sealed class Mov32Propagate : BaseTransform
+	public sealed class Mov32Coalescing : BaseTransform
 	{
-		public Mov32Propagate() : base(X64.Mov32, TransformType.Manual | TransformType.Optimization)
+		public Mov32Coalescing() : base(X64.Mov32, TransformType.Manual | TransformType.Optimization)
 		{
 		}
