+++ Source/Mosa.Platform.x86/Transforms/Optimizations/Auto/AutoTransforms.cs
@@ -13,7 +13,7 @@ namespace Mosa.Platform.x86.Transforms.Optimizations.Auto
 	public static class AutoTransforms
 	{
 		public static readonly List<BaseTransform> List = new List<BaseTransform> {
-			 new Standard.Mov32Consolidate(),
+			 new Standard.Mov32Coalescing(),
 			 new StrengthReduction.Add32ByZero(),
 			 new StrengthReduction.Sub32ByZero(),
 			 new Shrink.And32ByFF(),
