+++ Source/Mosa.Platform.x86/Transforms/Optimizations/Auto/Standard/Mov32Coalescing.cs
@@ -2,16 +2,18 @@
 
 // This code was generated by an automated template.
 
+using Mosa.Platform.x86;
 using Mosa.Compiler.Framework;
+using Mosa.Compiler.Framework.Transforms;
 
-namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
+namespace Mosa.Platform.x86.Transforms.Optimizations.Auto.Standard
 {
 	/// <summary>
-	/// Move32Propagation
+	/// Mov32Coalescing
 	/// </summary>
-	public sealed class Move32Propagation : BaseTransform
+	public sealed class Mov32Coalescing : BaseTransform
 	{
-		public Move32Propagation() : base(IRInstruction.Move32, TransformType.Auto | TransformType.Optimization)
+		public Mov32Coalescing() : base(X86.Mov32, TransformType.Auto | TransformType.Optimization)
 		{
 		}
 
@@ -23,10 +25,10 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 			if (context.Operand1.Definitions.Count != 1)
 				return false;
 
-			if (context.Operand1.Definitions[0].Instruction != IRInstruction.Move32)
+			if (context.Operand1.Definitions[0].Instruction != X86.Mov32)
 				return false;
 
-			if (IsCPURegister(context.Operand1.Definitions[0].Operand1))
+			if (!IsVirtualRegister(context.Operand1.Definitions[0].Operand1))
 				return false;
 
 			return true;
@@ -38,7 +40,7 @@ namespace Mosa.Compiler.Framework.Transforms.Optimizations.Auto.Simplification
 
 			var t1 = context.Operand1.Definitions[0].Operand1;
 
-			context.SetInstruction(IRInstruction.Move32, result, t1);
+			context.SetInstruction(X86.Mov32, result, t1);
 		}
 	}
 }
